Centralize run conclusion labels
